requirements.txt
adjustText==0.7.3
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
cycler==0.11.0
fonttools==4.29.1
google-api-core==2.5.0
google-api-python-client==2.38.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.55.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.2
oauthlib==3.2.0
outcome==1.1.0
packaging==21.3
Pillow==9.0.1
pkg_resources==0.0.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
RPi.GPIO==0.7.1
rsa==4.8
selenium==4.1.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
spidev==3.5
trio==0.20.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.8
wsproto==1.0.0
